requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
python-dotenv
Werkzeug
email-validator
Flask-Mail
SQLAlchemy